+++ src/main/java/com/zb/cinema/movie/model/ResponseMessageHeader.java
@@ -1,4 +1,4 @@
-package com.zb.cinema.admin.movie.model;
+package com.zb.cinema.movie.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
